requirements.txt
#########################

#! JARVIS Main System

#########################

cython==3.0.4
datetime==5.3
dlib==19.24.99
mediapipe==0.10.0
numpy==1.25.2
ollama==0.1.9
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.1.1
playsound==1.3.0
protobuf==3.20.3
pyautogui==0.9.54
pyaudio==0.2.13
pyttsx3==2.90
speechrecognition==3.10.1

#########################

#? JARVIS Plugins

#########################

anchormake==0.1.0
bleak==0.21.1
colorama==0.4.3
flask==3.0.0
requests==2.31.0
nanoleafapi==2.1.2
openai==0.27.8
openpyxl==3.1.2
paramiko==3.4.0
pupil-apriltags==1.0.4.post10
pymongo==4.6.1
speedtest==0.0.1
speedtest-cli==2.1.3
spotipy==2.23.0
wikipedia==1.4.0